Fixed linux & tracy

+++ font_rendering.h
@@ -1,8 +1,9 @@
 #include <GL/glew.h>
-#include <gl/GL.h>
+#include <GL/gl.h>
 #include <glm/glm.hpp>
 #include <map>
 #include <freetype/freetype.h>
+
 #include <iostream>
 
 struct Character {
@@ -73,7 +74,7 @@ int LoadFontRendering(std::string font_path) {
             font_texture,
             glm::ivec2(face->glyph->bitmap.width, face->glyph->bitmap.rows),
             glm::ivec2(face->glyph->bitmap_left, face->glyph->bitmap_top),
-            face->glyph->advance.x
+            static_cast<unsigned int>(face->glyph->advance.x)
         };
         Characters.insert(std::pair<char, Character>(c, character));
     }
